añadiendo sequelize de categorias

+++ database/models/Category.js
@@ -1,7 +1,7 @@
 module.exports = function(sequelize, dataTypes) {
     const Category = sequelize.define('categories', {
         id: {
-            type: dataTypes.INTEGER(10).UNSIGNED,
+            type: dataTypes.INTEGER.UNSIGNED,
             allowNull: false,
             primaryKey: true
         },
